install.sh: Don't overwrite "bashrc.cust.priv" by default
Copy it only if not existing or if '-f'/'--force' parameter is enabled.

+++ install.sh
@@ -10,5 +10,5 @@ sleep 3
 
 cp bashrc /etc/bashrc && chmod 644 /etc/bashrc
 cp bashrc.cust /etc/bashrc.cust && chmod 644 /etc/bashrc.cust
-cp bashrc.cust.priv /etc/bashrc.cust.priv && chmod 644 /etc/bashrc.cust.priv
+if [[ $1 = -f || $1 = --force  || ! -f /etc/bashrc.cust.priv ]]; then cp bashrc.cust.priv /etc/bashrc.cust.priv && chmod 644 /etc/bashrc.cust.priv; fi
 echo "Done."
